requirements.txt
python-dotenv
psycopg2-binary
sqlalchemy
databases[asyncpg]
databases[aiosqlite]
pydantic>=2.7.0
pydantic-settings
fastapi<=0.111.0
uvicorn
itsdangerous
line-bot-sdk<=3.12.0
openai<=1.43.0